requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
azure-identity==1.19.0
openai==1.51.2
python-multipart==0.0.9
gunicorn==23.0.0